* feature/cmake-update: (27 commits)
  added orUtils
  arc headers update
  change pricersolverwrapper.cpp
  change headers in files
  format simplification solver.h
  simplification header files arc
  simplification label.hpp
  newest version fmt
  simplification stabwrapper and name change pricerwrapper
  added config cmake or_utils
  format interval.h
  deleted or_utils headers from SOurcesandHeaders
  format util.h
  format forward BDD
  deleted not needed headers
  shorter definitions pricerevaluate
  removed model interface backward and forward bdd
  time_bis mv to or_utils
  mv binomial-heap to or_utils
  mv heap to or_utils
  ...

+++ include/util/DataTable.hpp
@@ -26,6 +26,7 @@
 #define DATA_TABLE_HPP
 
 #include <cassert>
+#include <cstddef>
 #include <ostream>
 
 #include "MyVector.hpp"
@@ -169,14 +170,14 @@ class DataTable {
      * @param i row index.
      * @return vector of elements on the row.
      */
-    MyVector<T>& operator[](int i) { return table[i]; }
+    MyVector<T>& operator[](size_t i) { return table[i]; }
 
     /**
      * Accesses to a row.
      * @param i row index.
      * @return vector of elements on the row.
      */
-    MyVector<T> const& operator[](int i) const { return table[i]; }
+    MyVector<T> const& operator[](size_t i) const { return table[i]; }
 
     friend std::ostream& operator<<(std::ostream& os, DataTable const& o) {
         for (int i = 0; i < o.numRows(); ++i) {
